refactored examples
